PCWEB-8492 build / prepublish 스크립트 분리
