+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@remember_web/ui",
-  "version": "0.0.0-beta.2",
+  "version": "0.0.0-beta.6",
   "description": "Remember UI Components",
   "homepage": "https://github.com/dramancompany/rui",
   "author": "Remember",
@@ -19,6 +19,8 @@
     "dev": "vite",
     "build": "rm -rf dist && tsc && node scripts/build.mjs",
     "build:visualize": "yarn build --visualize",
+    "pub": "yarn build && node scripts/prepublish.mjs && cd dist && npm publish",
+    "pub:beta": "yarn pub --tag beta",
     "lint": "eslint . --fix --ext .js,.jsx,.ts,.tsx",
     "prettier": "prettier -c src",
     "prettier-fix": "prettier --write src",
