+++ scripts/build.mjs
@@ -43,20 +43,6 @@ const args = process.argv.slice(2);
   if (args.includes('--visualize')) {
     open(path.resolve(__dirname, '..', 'visualizer', 'index.stats.html'));
   }
-
-  console.log('Copy files for distribution...');
-
-  const packageJsonObject = generatePackageJsonForDist();
-  const outDir = getOutDir();
-
-  fs.writeFileSync(
-    path.resolve(__dirname, '..', outDir, 'package.json'),
-    Buffer.from(JSON.stringify(packageJsonObject, null, 2), 'utf-8')
-  );
-  fs.copyFileSync(
-    path.resolve(__dirname, '..', 'README.md'),
-    path.resolve(__dirname, '..', outDir, 'README.md')
-  );
 })();
 
 function getModulesToBuild() {
@@ -83,30 +69,4 @@ function getOutputFileName(module) {
 
     return fileNameTokens.join('.');
   };
-}
-
-function generatePackageJsonForDist() {
-  const packageJson = fs.readFileSync(
-    path.resolve(__dirname, '../package.json').toString('utf-8')
-  );
-  const packageJsonObject = JSON.parse(packageJson);
-  packageJsonObject.scripts = {};
-  packageJsonObject.devDependencies = {};
-  if (packageJsonObject.main.startsWith('dist')) {
-    packageJsonObject.main = packageJsonObject.main.replace('dist/', '');
-  }
-  if (packageJsonObject.module.startsWith('dist')) {
-    packageJsonObject.module = packageJsonObject.module.replace('dist/', '');
-  }
-
-  return packageJsonObject;
-}
-
-function getOutDir() {
-  const tsconfigJson = fs.readFileSync(
-    path.resolve(__dirname, '../tsconfig.json').toString('utf-8')
-  );
-  const tsconfigJsonObject = JSON.parse(tsconfigJson);
-
-  return tsconfigJsonObject.compilerOptions.outDir.replace('./', '');
 }
